refactor python hook
